+++ classes/Markdownify/Converter.php
@@ -388,7 +388,8 @@ class Converter
 					}
 					break;
 				default:
-					trigger_error('invalid node type', E_USER_ERROR);
+					// FJ fix PHP8.4 deprecated passing E_USER_ERROR to trigger_error()
+					throw new \Exception('invalid node type');
 					break;
 			}
 			$this->lastWasBlockTag = $this->parser->nodeType == 'tag' && $this->parser->isStartTag && $this->parser->isBlockElement;
@@ -1091,7 +1092,8 @@ class Converter
 	protected function unstack()
 	{
 		if (!isset($this->stack[$this->parser->tagName]) || !is_array($this->stack[$this->parser->tagName])) {
-			trigger_error('Trying to unstack from empty stack. This must not happen.', E_USER_ERROR);
+			// FJ fix PHP8.4 deprecated passing E_USER_ERROR to trigger_error()
+			throw new \Exception('Trying to unstack from empty stack. This must not happen.');
 		}
 
 		return array_pop($this->stack[$this->parser->tagName]);